Fix how max-in-flight is tracked
No longer rely on the informer lister to calculate the number of
in-flight jobs, since that is eventually consistent

Instead introduce the limiter, which keeps track of how many jobs have
been passed to the scheduler and receives job completion events from an
informer.

This way we know about the job creations as they are created, instead of
there being some delay. We only use the informer for completions.