Turn some `require`s into `asserts`
`require` will stop the test on failure. This will hide subsequent
failure and we want to see as many failures as possible. It should only
be used if the failure is in some setup step and the continuing the test
does not make sense, for example, because it would cause nil pointer
dereferencing. In particular, using `require` in a loop is rarely
necessary.