Ignore strict host checking to prevent prompt from stalling test run
Without `StrictHostKeyChecking=no`, there can be a prompt to check the
fingerprints of the hosts. This test was only about whether ssh does not
run into permissions issues reading the .ssh directory. So, whether or
not the host is recognized is irrelevant.

In general, it is a not a good idea to use `StrictHostKeyChecking=yes`
(the default) when invoking ssh in a test when you can't guarantee that
the private key on the other end is stable. In this case, the other end
is github.com, so that guarantee is very far from being met.