Remove `-G.` from code generation check
This was in place because agent-stack-k8s used to mess with the file
permissions in the workspace. That's no longer the case, so we can
remove it. `-G.` effectively means ignore permissions in the diff, but I
think its safer to not ignore them as long as they don't cause false
positives, which they always did prior to https://github.com/buildkite/agent-stack-k8s/pull/246