Use DefaultAgentImage for the BuildFailureJob
Previously, the agent image configured in the controller's config was
used for the BuildFailureJob. But customers could have configured this
to be a private image. We don't want to copy the imagePullSecrets in as
there could be an error in their formatting, and the failure job is
designed to bubble up such errors to the job logs. So we use the default
agent image which is guaranteed to be public.