Fix `pod-spec-patch` was not unmarshalled correctly by the controller config
The `corev1.PodSpec` type has (nested) fields that have a struct tag
`json:",inline"`. This means they should be ignored when unmarshalling,
but their child field should take their place.

In particular, the `SecretRef` type, which we use extensively, has a
inline field `corev1.LocalObjectReference`. As this was not being
unmarshalled correctly, users that specified the following pod-spec-patch
in their config would see errors:
```yaml
config:
  pod-spec-patch:
    containers:
    - name: checkout
      envFrom:
      - secretRef:
          name: git-ssh-secrets
```
The controller would only to see the `name` field in the `secretRef` as
an empty string. This manifested as a runtime error when the controller
attempts to create jobs in k8s, it would complain that the empty string
was not a valid secret name.

This bug occur because, viper does not handle this situation correctly.
The reason it does not is because it unmarshalls into a map first, and
then decodes the map using `mapstructure`. In `mapstructure`, the
`inline` tag is not recognized, but a `squash` tag provides the same
functionality. So I've made a fork of `mapstructure` to equate `inline`
with `squash`. This fixes the issue.

Long term, we should upstream a fix to either `viper` or `mapstructure`,
but for now, this is good enough.