Ensure all agent tags are present in job tags
The shared informer watches Jobs matching Labels created from tags
in the controller's config. When a scheduled job is retrieved it was
previously checking if all job tags were present as agent tags (and
not the other way around). This was allowing scheduled jobs with only
a matching `queue` tag to be created as a k8s Job. This then prevented
the informer from matching the Labels of these Jobs and the controller
was unable to know when they finished.